% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{alg_deg}
\alias{alg_deg}
\title{Algebraic degree of semidefinite programming}
\usage{
alg_deg(n)
}
\arguments{
\item{n}{size of matrix}
}
\value{
The output of \code{alg_deg} is a table of the algebraic degrees
        for the given value of \code{n}. To avoid loss of precision,
        the algebraic degrees are given as strings. See the below example for
        a simple conversion to numeric, including the problems associated with
        that step.
}
\description{
\code{alg_deg} produces a table for the algebraic degree of semidefinite
programmin (see the accompanying vignette for reference to literature).
}
\details{
Only some precomputed values of \code{n} are available. See the literature
referenced in the accompanying vignette for general (combinatorial) formulas
for the algebraic degree.
}
\section{See also}{

\code{\link[symconivol]{curv_meas_exact}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
AD_str <- alg_deg(6)

# convert to integer matrix
AD <- matrix( as.numeric(AD_str), dim(AD_str) )
colnames(AD) <- colnames(AD_str)
rownames(AD) <- rownames(AD_str)

# compare both matrices
print(AD_str)
print(AD)

# doing the same for larger n
AD_str <- alg_deg(14)
AD <- matrix( as.numeric(AD_str), dim(AD_str) )

# the conversion introduces rounding errors
i <- which.max(AD_str)
print(AD_str[i])
print(AD[i],digits=20)

}
