% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alg_deg_data}
\alias{alg_deg_data}
\title{Algebraic degree of semidefinite programming}
\format{A list of 13 character matrices; the rows and columns corresponds to
m and r, respectively.}
\source{
The degrees have been computed with Maple using a formula given in
\url{https://arxiv.org/abs/1509.05241}, \url{https://projecteuclid.org/euclid.kmj/1478073765}.
}
\usage{
alg_deg_data
}
\description{
A list of the values of the algebraic degree of semidefinite programming
delta(m,n,r) for \code{n=2,3,...,14}.
The values are given as strings to avoid rounding errors.
}
\section{See also}{

\code{\link[symconivol]{alg_deg}}

Package: \code{\link[symconivol]{symconivol}}
}

\keyword{datasets}
