% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constr_eigval.R
\name{constr_eigval}
\alias{constr_eigval}
\title{Stan model creation for sampling index constrained eigenvalues from the Gaussian orthogonal/unitary/symplectic ensembles}
\usage{
constr_eigval(pos, free, neg, filename = NA, overwrite = FALSE)
}
\arguments{
\item{pos}{parameter description}

\item{free}{parameter description}

\item{neg}{parameter description}

\item{filename}{parameter description}

\item{overwrite}{parameter description}
}
\value{
returns what
}
\description{
\code{constr_eigval}
}
\section{See also}{

\code{\link[symconivol]{constr_eigval_to_bcbsq}}, 
\code{\link[symconivol]{prepare_em_cm}}, 
\code{\link[symconivol]{estim_em_cm}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
# some example code

}
