% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_curvmeas.R
\name{constr_eigval_to_bcbsq}
\alias{constr_eigval_to_bcbsq}
\title{Converting samples of index constrained eigenvalues to samples of bivariate-chi-bar-squared distribution}
\usage{
constr_eigval_to_bcbsq(pos, free, neg, samp = NA, filename = NA)
}
\arguments{
\item{pos}{parameter description}

\item{free}{parameter description}

\item{neg}{parameter description}

\item{samp}{parameter description}

\item{filename}{parameter description}
}
\value{
returns what
}
\description{
\code{constr_eigval_to_bcbsq}
}
\section{See also}{

\code{\link[symconivol]{constr_eigval}}, 
\code{\link[symconivol]{prepare_em_cm}}, 
\code{\link[symconivol]{estim_em_cm}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
# some example code

}
