% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{curv_meas_exact}
\alias{curv_meas_exact}
\title{Exact curvature measures of symmetric cones}
\usage{
curv_meas_exact(beta, n)
}
\arguments{
\item{beta}{Dyson index specifying the underlying (skew-) field:
\describe{
  \item{\code{beta==1}:}{real numbers}
  \item{\code{beta==2}:}{complex numbers}
  \item{\code{beta==4}:}{quaternion numbers}
}}

\item{n}{size of matrix.}
}
\value{
The output of \code{curv_meas_exact} is a list of six elements:
        \itemize{
          \item \code{A}: the combined matrix of curvature measures; \code{A}
              is given in terms of the other parameters by
              \code{(A_const + A_sqrt2*sqrt(2) + A_piinv/pi + A_sqrt2_pi*sqrt(2)/pi)/denom}
          \item \code{A_const}: integer matrix for the constant term
          \item \code{A_sqrt2}: integer matrix for the \code{sqrt(2)} term
          \item \code{A_piinv}: integer matrix for the \code{1/pi} term
          \item \code{A_sqrt2_pi}: integer matrix for the \code{sqrt(2)/pi} term
          \item \code{denom}: common denominator of all terms
        }
}
\description{
\code{curv_meas_exact} returns the exact values of the curvature measures of
symmetric cones, which are known (\code{n==1,2,3}).
}
\details{
The known curvature measures are elements in the ring \code{Q[sqrt(2),1/pi]},
so the exact values can be given in terms of integers corresponding to a 
common denominator and corresponding integer matrices for the coefficients
of the natural expansion in \code{1, sqrt(2), 1/pi, sqrt(2)/pi}. These matrices
are returned by this function, along with the denominator and the combined
matrix of curvature measures.
}
\section{See also}{

\code{\link[symconivol]{alg_deg}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
# considering the case of 3x3 complex unitary matrices
CM <- curv_meas_exact(2,3)

# sum of intrinsic volumes is equal to one
sum( CM$A )

# sum of even (and odd) index intrinsic volumes is 1/2
sum( CM$A \%*\% rep_len(c(1,0),dim(CM$A)[2]) )

# A is given by combining the remaining matrices and the denominator
norm( CM$A - ( CM$A_const + CM$A_sqrt2*sqrt(2) + CM$A_piinv/pi + CM$A_sqrt2_pi*sqrt(2)/pi )/CM$denom )

}
