% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_curvmeas.R
\name{estim_em_cm}
\alias{estim_em_cm}
\title{Estimating curvature measures from bivariate chi-bar-squared data using EM algorithm}
\usage{
estim_em_cm(d, low, upp, m_samp, N = 20, no_of_lcc_projections = 1,
  data = NULL)
}
\arguments{
\item{d}{parameter description}

\item{low}{parameter description}

\item{upp}{parameter description}

\item{m_samp}{parameter description}

\item{N}{parameter description}

\item{data}{parameter description}

\item{no_off_lcc_projections}{parameter description}
}
\value{
returns what
}
\description{
\code{estim_em_cm}
}
\section{See also}{

\code{\link[symconivol]{constr_eigval}}, 
\code{\link[symconivol]{constr_eigval_to_bcbsq}}, 
\code{\link[symconivol]{prepare_em_cm}},
\code{\link[symconivol]{indnorm_to_unnorm}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
# some example code

}
