% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_prob}
\alias{ind_prob}
\title{Ratios of index probabilities}
\format{A list of two-element integer vectors.}
\source{
The samples have been found with the HMC sampler Stan with a model
that simulates the eigenvalues of a Gaussian orthogonal/unitary/symplectic
ensemble with index constrained to \{r,r+1\}. The code below shows how this data
can be generated.
}
\usage{
ind_prob
}
\description{
A list of sample counts of a Bernoulli variable with (unnormalized) success
and failure probabilities given by Prob\{ind=r\} and Prob\{ind=r+1\}.
}
\details{
The values of the parameters beta, n, r are specified in the names of the list
elements, which are of the form \code{beta=_,n=_,r=_} with _ replaced by
values. The list contains all combinations of \code{beta=1,2,4}, \code{n=4,5,...,10},
and \code{r} between \code{floor((n+1)/2)} and \code{n-2}. The remaining values
can be found through the relation Prob\{ind=r\}=Prob\{ind=n-r\}.
}
\section{See also}{

\code{\link[symconivol]{constr_eigval}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
\dontrun{
library(rstan)

ind_prob <- list()
nmin <- 4
nmax <- 10
filename <- "tmp.stan"
N <- 1e6
warmup <- 1e3
for (n in nmin:nmax) {
    for (beta in c(1,2,4)) {
        for (r in floor((n+1)/2):(n-2)) {
            index <- str_c("beta=",beta,",n=",n,",r=",r)
            M <- constr_eigval(beta, n, r, r+1, filename=filename)
            stan_samp <- stan( file = filename, data = M$data,
                               chains = 1, warmup = warmup,
                               iter = N+warmup, cores = 2, refresh = 1e4 )
            ef <- rstan::extract(stan_samp)$ef
            indprob[[index]] <- c(length(which(ef<0)),length(which(ef>0)))
        }
    }
}
file.remove(filename)
}

}
\keyword{datasets}
