% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indnorm_to_unnorm.R
\name{indnorm_to_unnorm}
\alias{indnorm_to_unnorm}
\title{Estimating curvature measures from index normalized curvature measures}
\usage{
indnorm_to_unnorm()
}
\arguments{
\item{d}{parameter description}
}
\value{
returns what
}
\description{
\code{indnorm_to_unnorm}
}
\section{See also}{

\code{\link[symconivol]{estim_em_cm}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
# some example code

}
