% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{leigh}
\alias{leigh}
\title{Leigh's curve}
\usage{
leigh(N = 1000)
}
\arguments{
\item{N}{number of intermediate points; size of resulting table}
}
\value{
The output of \code{leigh} is a list of three elements:
        \itemize{
          \item \code{tab}: a table of function values with \code{N} elements
          \item \code{lkup_rho}: a corresponding lookup function for rho in terms of kappa
          \item \code{lkup_kappa}: a corresponding lookup function for kappa in terms of rho
        }
}
\description{
\code{leigh} produces a table and lookup functions for Leigh's curve
(see accompanying vignette for definition).
}
\section{See also}{

\code{\link[symconivol]{pat_bnd}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
L <- leigh()
ggplot(L$tab, aes(x=rho,y=kappa)) + geom_line()
x <- (0:10)/10
matrix(c(x, L$lkup_rho(x) ),11,2)
matrix(c(x, L$lkup_kappa(x) ),11,2)

}
