% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mu}
\alias{mu}
\title{Estimated limit curve for the dimension normalized curvature measures}
\usage{
mu()
}
\value{
The output of \code{mu} is a list of three elements:
        \itemize{
          \item \code{table}: a table of function values
          \item \code{lkup_rho}: a corresponding lookup function for rho in terms of kappa
          \item \code{lkup_kappa}: a corresponding lookup function for kappa in terms of rho
        }
}
\description{
\code{mu} returns a pre-computed table and lookup functions for the 
estimated limit curve for dimension normalized curvature measures
(see accompanying vignette for definition; we use \code{C=0.2}).
}
\section{See also}{

\code{\link[symconivol]{leigh}}, \code{\link[symconivol]{rate}}, 
\code{\link[symconivol]{pat_bnd}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
M <- mu()
ggplot(M$table, aes(x=rho,y=kappa)) + geom_line()
rho <- (0:10)/10
matrix(c(rho, M$lkup_kappa(rho)),11,2)
kappa <- (0:10)/10
matrix(c(M$lkup_rho(kappa), kappa),11,2)

}
