% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mu_data}
\alias{mu_data}
\title{Values of estimated limit curve of dimension normalized curvature measures}
\format{A two-column table of function values of mu; the columns corresponds to
\code{rho} and \code{kappa=mu(rho)}.}
\source{
The function values have been computed with an estimated value of C=0.2.
The code below shows how this data can be generated.
}
\usage{
mu_data
}
\description{
A table of function values of the estimated limit curve of
dimension normalized curvature measures mu.
}
\section{See also}{

\code{\link[symconivol]{mu}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
\dontrun{
library(tidyverse)

L <- leigh(1e5)
R <- rate(1e5)
rho <- seq(0,1,length.out=1e4)
Lkappa <- L$lkup_kappa(rho)
Rkappa <- R$lkup_R(rho)
kappa <- seq(0,1,length.out=1e4)
C <- 0.2
rhomin <- sapply(kappa, function(x)
    return( max(1-sqrt(1-x), min(sqrt(x),rho[which.min(
        (x-Lkappa)^2 / (1-abs(2*rho-1)) + 4*C*Rkappa )] ))))

mu_data <- tibble(kappa=kappa,rho=rhomin)
}

}
\keyword{datasets}
