% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pat_bnd}
\alias{pat_bnd}
\title{Pataki bounds}
\usage{
pat_bnd(beta, n)
}
\arguments{
\item{beta}{Dyson index specifying the underlying (skew-) field:
\describe{
  \item{\code{beta==1}:}{real numbers}
  \item{\code{beta==2}:}{complex numbers}
  \item{\code{beta==4}:}{quaternion numbers}
}}

\item{n}{size of matrix}
}
\value{
The output of \code{pat_bnd} is a list of five elements:
        \itemize{
          \item \code{d}: ambient dimension of corresponding Euclidean space,
              given in terms of \code{beta,n} by \code{n+beta*choose(n,2)}
          \item \code{k_low}: function returning the lower bound for index \code{k}, given \code{r}
          \item \code{k_upp}: function returning the upper bound for index \code{k}, given \code{r}
          \item \code{r_low}: function returning the lower bound for index \code{r}, given \code{k}
          \item \code{r_upp}: function returning the upper bound for index \code{r}, given \code{k}
        }
}
\description{
\code{pat_bnd} provides the Pataki inequalities for given \code{beta} and \code{n}.
}
\section{See also}{

\code{\link[symconivol]{leigh}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
# considering the case of 3x3 complex unitary matrices
beta <- 2
n <- 3
P <- pat_bnd(beta,n)

# ambient dimension
P$d

# bounds for index k, given r
P$k_low(0:n)
P$k_upp(0:n)

# bounds for index r, given k
P$r_low(0:P$d)
P$r_upp(0:P$d)

}
