% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rate}
\alias{rate}
\title{Large deviation rate function for index probability}
\usage{
rate(N = 1000)
}
\arguments{
\item{N}{number of intermediate points; size of resulting table}
}
\value{
The output of \code{rate} is a list of four elements:
        \itemize{
          \item \code{table}: a table of function values with \code{N} elements
          \item \code{table}: a table of function values with \code{N} elements for the derivative
          \item \code{lkup_R}: a corresponding lookup function for the rate function R
          \item \code{lkup_dR}: a corresponding lookup functions for the derivative of R
        }
}
\description{
\code{rate} produces a table and lookup function for the large deviation rate function
of the index (see accompanying vignette for definition).
}
\section{See also}{

\code{\link[symconivol]{leigh}}, \code{\link[symconivol]{mu}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
R <- rate()
ggplot(R$table, aes(x=rho,y=R)) + geom_line()
ggplot(R$dtable, aes(x=rho,y=dR)) + geom_line()
rho <- (0:10)/10
matrix(c(rho, R$lkup_R(rho), R$lkup_dR(rho) ),11,3)

}
