% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symconivol.R
\docType{package}
\name{symconivol}
\alias{symconivol}
\alias{symconivol-package}
\title{symconivol: A package for curvature measures of symmetric cones}
\description{
The symconivol package provides functions for ...
}
\details{
\code{symconivol} supports ... Stan.
}
\section{Functions}{

\itemize{
  \item \code{\link[symconivol]{constr_eigval}}: ...
  \item \code{\link[symconivol]{prepare_em_cm}}: ...
  \item \code{\link[symconivol]{estim_em_cm}}: ...
}
}

\section{See also}{

\describe{
  \item{github}{\href{https://github.com/damelunx/conivol}{github.com/damelunx/conivol}}
  \item{vignette}{\href{../doc/conic-intrinsic-volumes.html}{Conic intrinsic
        volumes and (bivariate) chi-bar-squared distribution}:
        introduces conic intrinsic volumes and (bivariate) chi-bar-squared distributions,
        as well as the computations involving polyhedral cones}
  \item{vignette}{\href{../doc/estim-conic-intrinsic-volumes-with-EM.html}{Estimating conic intrinsic volumes from bivariate chi-bar-squared data}:
        describes the details of the algorithm for finding the intrinsic volumes of closed
        convex cones from samples of the associated bivariate chi-bar-squared distribution}
  \item{vignette}{\href{../doc/bayesian.html}{Bayesian estimates for conic intrinsic volumes}:
        describes the Bayesian approach for reconstructing intrinsic volumes
        from sampling data, which can either be samples from the intrinsic
        volumes distribution (in the case of polyhedral cones), or from the
        bivariate chi-bar-squared distribution, and which can be with or without
        enforcing log-concavity of the intrinsic volumes}
}
}

