% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_curvmeas.R
\name{prepare_em_cm}
\alias{prepare_em_cm}
\title{Evaluate bivariate chi-bar-squared samples for maximum likelihood estimation}
\usage{
prepare_em_cm(d, low, upp, m_samp)
}
\arguments{
\item{d}{parameter description}

\item{low}{parameter description}

\item{upp}{parameter description}

\item{m_samp}{parameter description}
}
\value{
returns what
}
\description{
\code{prepare_em_cm}
}
\section{See also}{

\code{\link[symconivol]{constr_eigval}}, 
\code{\link[symconivol]{constr_eigval_to_bcbsq}}, 
\code{\link[symconivol]{estim_em_cm}}

Package: \code{\link[symconivol]{symconivol}}
}

\examples{
# some example code

}
